requirements.txt
discord
mysql-connector-python
python-dotenv
pillow
docker
paramiko
mcrcon